refactor(schemas): relocate per-host settings schemas to their host packages
The four-way settings split (Core + VS Code + Desktop + CLI) landed in
src/shared/schemas/settings/ as scaffolding. Moving the host-specific
schemas into the packages that own them so additions land naturally and
the shared layer only carries genuinely cross-host material.

Final layout:

  src/shared/schemas/coreSettings.ts                 — host-neutral
  packages/extension/src/schemas/vscodeSettings.ts   — VS Code-only
  packages/extension/src/schemas/texraSettings.ts    — Core + VS Code
                                                       composition,
                                                       flatten/build
                                                       helpers, JSON
                                                       schema generator
  packages/desktop/src/main/schemas/desktopSettings.ts — empty extension
                                                          placeholder
  packages/cli/src/schemas/cliSettings.ts            — CLI runtime fields
  packages/cli/src/schemas/knownKeys.ts              — CLI's
                                                       KNOWN_TEXRA_KEYS
                                                       derived from
                                                       CORE_SETTING_PATHS +
                                                       CLI_SETTING_PATHS

Other changes:

- src/utils/config/settingsSchema.ts deleted. The CLI is the only
  consumer of KNOWN_TEXRA_KEYS / CLI_APPROVAL_POLICIES /
  CLI_OUTPUT_FORMATS, so those move into the CLI package
  (cliSettings.ts + knownKeys.ts). approvalPolicy.ts and cliConfig.ts
  switch to relative imports.
- src/shared/schemas/settingsConfiguration.ts slimmed to a backward-
  compatible re-export shim that surfaces the Core schema for legacy
  consumers; the composed TexraSettings and tooling moved to the
  extension package.
- src/common/files/fileTypeUtils.ts and fileListingRules.ts switch
  from DEFAULT_TEXRA_SETTINGS to DEFAULT_CORE_SETTINGS (the .files.*
  defaults they read are Core, not VS Code-specific).
- CORE_SETTING_PATHS exported from coreSettings.ts and
  VSCODE_SETTING_PATHS from vscodeSettings.ts so TEXRA_SETTING_PATHS
  becomes [...CORE_SETTING_PATHS, ...VSCODE_SETTING_PATHS] instead of
  hand-maintaining a flat list.
- New @extensionSchemas/* path alias points at
  packages/extension/src/schemas/* for the test files and other shared
  consumers that need the composed VS Code schema.
- scripts/sync-settings-configuration.mjs updated to load
  buildTexraPackageConfiguration from the new compiled location.
- Mocha and vitest settings tests updated to the new import paths.

Net: -1007 / +30 lines (mostly the duplicate path list that became a
composition; per-host directories make settings ownership explicit).
All 898 vitest tests pass; full typecheck (root + extension + cli +
desktop) clean; 0 lint errors.
