+++ packages/cli/src/runtime/approvalPolicy.ts
@@ -1,4 +1,4 @@
 export {
   CLI_APPROVAL_POLICIES,
   type CliApprovalPolicy,
-} from '@utils/config/settingsSchema';
+} from '../schemas/cliSettings';
