+++ packages/cli/src/schemas/cliSettings.ts
