+++ packages/desktop/src/main/schemas/desktopSettings.ts
