+++ src/utils/config/index.ts
@@ -1,3 +1,2 @@
 export * from './configUtils';
 export * from './constants';
-export * from './settingsSchema';
